Move `GitRepoMetricAggregate` to `join` subpackage
